Rename code_ into execute_code and update tutorial
