+++ docs/sphinx_doc/source/agentscope.service.execute_code.rst
@@ -4,7 +4,7 @@ Code package
 exec\_python module
 --------------------------------------------
 
-.. automodule:: agentscope.service.code.exec_python
+.. automodule:: agentscope.service.execute_code.exec_python
    :members:
    :undoc-members:
    :show-inheritance:
