+++ src/agentscope/service/__init__.py
@@ -2,7 +2,7 @@
 """ Import all service-related modules in the package."""
 from loguru import logger
 
-from .code_.exec_python import execute_python_code
+from .execute_code.exec_python import execute_python_code
 from .file.common import (
     create_file,
     delete_file,
